# perf(client): prealloc pkg/varlog.(*transmitter).transmitQueue