@ijsong 2024-03-05
### Merge activity

* **Mar 5, 4:41 AM EST**: @ijsong started a stack merge that includes this pull request via [Graphite](https://app.graphite.dev/github/pr/kakao/varlog/721).